updates to push schema changes to azure

+++ MLDev.LOTOW/Models/Stat.cs
@@ -6,7 +6,7 @@ namespace MLDev.LOTOW.Models
     public class Stat
     {
         public int StatId { get; set; }
-        public decimal Order { get; set; }
+        public int Order { get; set; }
         public string Name { get; set; }
         public bool IsFinite { get; set; }// if this stat has a set ammount
     }
